+++ module.json
@@ -2,7 +2,7 @@
   "id": "visage",
   "title": "Visage",
   "description": "Allow token art and actor portraits to be easily switched to alternative images.",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "compatibility": {
     "minimum": "13",
     "verified": "13"
@@ -13,8 +13,7 @@
     }
   ],
   "esmodules": [
-    "main.js",
-    "visage-selector.js"
+    "main.js"
   ],
   "styles": [
     "styles/visage.css"
